better frontend+bridge interaction with c
- better bridge mimic in bridge/__init__.py (lazy loading)
- bridge state enforces proposed interpretation when getting/setting elements
- corrected sim/bridge dependancies in build (added paths.py to both and added bridge.h to sim)
- added -c directive detection to build

+++ sim/bruv/paths.py
@@ -10,22 +10,26 @@ from pathlib import Path
 
 
 ROOT = Path(__file__).resolve().parent.parent
-def fromroot(rel):
-    return ROOT / rel
 
 
 def shortstr(path):
+    path = Path(path).resolve()
     if path.is_relative_to(ROOT):
         path = path.relative_to(ROOT)
-    return path.as_posix()
+    return repr(path.as_posix())
 
 
-def subfiles(directory, ext="", as_str=False, as_rel=False):
-    files = [p for p in directory.glob("**/*" + ext) if p.is_file()]
+def subfiles(directory, ext="", recursive=True, as_str=False, as_rel=False):
+    directory = Path(directory)
+    pattern = "**/"*bool(recursive) + "*" + ext
+    files = [path for path in directory.glob(pattern) if path.is_file()]
+    parts = lambda p: p.relative_to(directory).parts
+    # Ignore hidden.
+    dirs = [path for path in files
+            if all(not part.startswith(".") for part in parts(path))]
     # Ignore files inside pycaches.
-    files = [p for p in files if ("__pycache__" not in p.resolve().parts)]
-    # Ignore hidden files.
-    files = [p for p in files if not p.name.startswith(".")]
+    files = [path for path in files
+             if ("__pycache__" not in parts(path))]
     # Transform to requested format.
     if as_rel:
         files = [x.relative_to(directory) for x in files]
@@ -33,10 +37,14 @@ def subfiles(directory, ext="", as_str=False, as_rel=False):
         files = [str(x) for x in files]
     return files
 
-def subdirs(directory, as_str=False, as_rel=False):
-    dirs = [p for p in directory.glob("**/*") if p.is_dir()]
-    # Ignore hidden files.
-    dirs = [p for p in dirs if not p.name.startswith(".")]
+def subdirs(directory, recursive=True, as_str=False, as_rel=False):
+    directory = Path(directory)
+    pattern = "**/"*bool(recursive) + "*"
+    dirs = [path for path in directory.glob(pattern) if path.is_dir()]
+    parts = lambda p: p.relative_to(directory).parts
+    # Ignore hidden.
+    dirs = [path for path in dirs
+            if all(not part.startswith(".") for part in parts(path))]
     # Note we dont ignore pycaches.
     # Transform to requested format.
     if as_rel:
@@ -46,19 +54,30 @@ def subdirs(directory, as_str=False, as_rel=False):
     return dirs
 
 
-def wipe(directory, including_itself=True, ignore_errors=False):
-    if including_itself:
-        shutil.rmtree(directory, ignore_errors=ignore_errors)
+def wipe(directory, missing_ok=False, only_contents=False):
+    directory = Path(directory)
+    if not directory.exists():
+        if not missing_ok:
+            raise FileNotFoundError("cannot wipe missing directory: "
+                                   f"{shortstr(directory)}")
         return
-    for entry in directory.iterdir():
-        if entry.is_dir():
-            shutil.rmtree(entry, ignore_errors=ignore_errors)
-        else:
-            entry.unlink()
+    if not directory.is_dir():
+        raise NotADirectoryError("cannot wipe non-directory: "
+                                f"{shortstr(directory)}")
+    if not only_contents:
+        shutil.rmtree(directory)
+    else:
+        for entry in directory.iterdir():
+            if entry.is_dir():
+                shutil.rmtree(entry)
+            else:
+                entry.unlink()
 
 
 def max_mtime(files):
+    files = [Path(p) for p in files]
     return max([(-1 if not p.is_file() else p.stat().st_mtime) for p in files])
+
 
 class pushd:
     """
@@ -70,7 +89,7 @@ class pushd:
         # Default to repo root.
         if path is None:
             path = ROOT
-        self.path = path
+        self.path = Path(path).resolve()
 
     def __call__(self, func):
         @functools.wraps(func)
@@ -81,8 +100,8 @@ class pushd:
         return wrapper
 
     def __enter__(self):
-        self._old_dir = Path.cwd()
-        os.chdir(self.path)
+        self._old_dir = os.getcwd()
+        os.chdir(str(self.path))
     def __exit__(self, etype, evalue, etb):
         os.chdir(self._old_dir)
 
@@ -91,17 +110,19 @@ GCC = "gcc"
 # if gcc isnt on path, you may edit this to correctly point to the executable.
 
 
-SIM = fromroot("bruv/sim")
+BRUV = ROOT / "bruv"
 
-BRIDGE = fromroot("bruv/bridge")
+SIM = BRUV / "sim"
 
-BIN = fromroot("bin")
+BRIDGE = BRUV / "bridge"
+
+BIN = ROOT / "bin"
 BIN_SIM = BIN / "sim"
 BIN_BRIDGE = BIN / "bridge"
 
-SIM_OBJ    = BIN_SIM / "sim.o"
-SIM_DISAS  = BIN_SIM / "sim.s"
 SIM_PREPRO = BIN_SIM / "sim.i"
+SIM_DISAS  = BIN_SIM / "sim.s"
+SIM_OBJ    = BIN_SIM / "sim.o"
 
 SIM_LIB_NAME = "sim"
 if sys.platform == "win32":
@@ -114,5 +135,46 @@ else:
     SIM_LIB = BIN_SIM / f"lib{SIM_LIB_NAME}.so"
     SIM_STUBS = None
 
+BRIDGE_MODULE_NAME = "bridge_cythonised"
+
 SIM_CACHE = BIN_SIM / "cache.json"
 BRIDGE_CACHE = BIN_BRIDGE / "cache.json"
+
+PATHS_PY = BRUV / "paths.py"
+BUILD_PY = BRUV / "build.py"
+BRIDGE_H = BRIDGE / "bridge.h"
+
+
+def sim_deps():
+    paths = subfiles(SIM)
+    paths.append(BUILD_PY) # always depends on builder.
+    paths.append(PATHS_PY) # builder depends on paths.
+    paths.append(BRIDGE_H) # depends on bridge.h also.
+    for p in paths:
+        if not p.is_file():
+            raise FileNotFoundError(f"missing sim dependancy: {shortstr(p)}")
+    return paths
+
+def bridge_deps():
+    paths = subfiles(BRIDGE)
+    paths.append(BUILD_PY) # always depends on builder.
+    paths.append(PATHS_PY) # builder depends on paths.
+    for p in paths:
+        if not p.is_file():
+            raise FileNotFoundError(f"missing bridge dependancy: {shortstr(p)}")
+    return paths
+
+
+def sim_built(): # not necessarily all existent.
+    paths = [SIM_CACHE]
+    paths.append(SIM_LIB)
+    if SIM_STUBS is not None:
+        paths.append(SIM_STUBS)
+    return paths
+
+def bridge_built(): # not necessarily all existent.
+    paths = [BRIDGE_CACHE]
+    # Finding the "correct" compiled bridge file exists is lowk hard, so instead
+    # we just check for any .pyd in the right spot.
+    paths += subfiles(BIN_BRIDGE, ext=".pyd", recursive=False)
+    return paths
